chore: Refactor recipe pages and components for better organization and reusability
