+++ src/pages/recipes/components/StrikeableStep.tsx
